Change KeyEvent.sequenceNumber to long from BigInteger
A long provides for ample size. A KEL with single byte events would be
measured in exabytes. I think we'll be okay and if not, we can change
this back on that wonderous day.

+++ core/src/main/java/foundation/identity/keri/KeyEventStore.java
@@ -25,7 +25,7 @@ public interface KeyEventStore {
 
   Stream<KeyEvent> streamKeyEvents(Identifier identifier);
 
-  Stream<KeyEvent> streamKeyEvents(Identifier identifier, BigInteger from);
+  Stream<KeyEvent> streamKeyEvents(Identifier identifier, long from);
 
   Optional<KeyState> getKeyState(Identifier identifier);
